Moved levelup settings to settings.js
